wsl-integration: Fix unbound variable errors
I agree that `set -eu` is good practice, but in this case this behavior
is being inherited for the user's bash session as a whole, which is not
expected. Adding `set +eu` at end of file to bring back expected
behavior.

Also adding UBUNTU_WSL variable checks, if they are not defined,
default values will be set and the user will be notified, by echo, that
the `ubuntu-wls.conf` is not up to date.

Signed-off-by: Matheus Castello <matheus@castello.eng.br>